% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_plot_functions.R
\name{grada_plot}
\alias{grada_plot}
\title{GRADA - Plot}
\usage{
grada_plot()
}
\value{
R barplots of the adapter positions.
}
\description{
This function will refer to the standard plot function.
}
\details{
It will only accept standard settings. You cannot define anything here. Please use another _plot function, if you want to change the in-/output.

Standard plot function: grada_plot_bar()
}
