% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_plot_functions.R
\name{grada_plot_bar}
\alias{grada_plot_bar}
\title{GRADA - Barplot}
\usage{
grada_plot_bar(
  PE = TRUE,
  input = "temp/",
  M_min = 0,
  M_max = 0,
  skip = TRUE,
  plot_row = 2,
  plot_col = 2
)
}
\arguments{
\item{PE}{paired data? TRUE / FALSE (std. TRUE)}

\item{input}{input folder where the "adapter_Positions.Rdata" is (output from grada_analyze_positions() std. "temp/")}

\item{M_min}{min mismatches. (not more than 5 plots allowed)}

\item{M_max}{max mismatches. (not more than 5 plots allowed)}

\item{skip}{if TRUE, it will skip plots for empty Data (if adapter is not found) so you will not get empty plots (sometimes it happens that the first plost is present anyway. this is due to a matrix R-command). (std. TRUE)}

\item{plot_row}{is the par(mfrow=c(plot_row,plot_col)) for arrenging the plots (std. 2)}

\item{plot_col}{is the par(mfrow=c(plot_row,plot_col)) for arrenging the plots (std. 2)}
}
\value{
R barplots of the adapter positions.
}
\description{
This function will plot the results of grada_analyze_positions() which are stored in "Adapter_Positions.Rdata"
}
\details{
The first position of the found Sequence will be plotted! If the read is 100 bases long and the adapter 10, the maximal position will therefore be 91!

plots will be skipped if no adapters are found. If there are any difficulties with this function, may set skip to FALSE.
}
