% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_plot_functions.R
\name{grada_plot_bar_full}
\alias{grada_plot_bar_full}
\title{GRADA - Barplot Full}
\usage{
grada_plot_bar_full(
  PE = TRUE,
  input = "temp/",
  skip = TRUE,
  plot_row = 2,
  plot_col = 2,
  colour = 2
)
}
\arguments{
\item{PE}{paired data? TRUE / FALSE (std. TRUE)}

\item{input}{input folder where the "adapter_Positions.Rdata" and "grada_table.txt" is (output from grada_analyze() / _positions() std. "temp/")}

\item{skip}{if TRUE, it will skip plots for empty Data (if adapter is not found) so you will not get empty plots (sometimes it happens that the first plost is present anyway. this is due to a matrix R-command). (std. TRUE)}

\item{plot_row}{is the par(mfrow=c(plot_row,plot_col)) for arrenging the plots (std. 2)}

\item{plot_col}{is the par(mfrow=c(plot_row,plot_col)) for arrenging the plots (std. 2)}

\item{colour}{colour scheme for plot. (1 = std. / 2 = colorblind / 3 = black and white)}
}
\value{
R barplots of the adapter positions.
}
\description{
This function will plot the results of grada_analyze_positions() which stored in "Adapter_Positions.Rdata"
}
\details{
All positions! of the found Sequence will be plotted.

plots will be skipped if no adapters are found. If there are any difficulties with this function, may set skip to FALSE.
}
