% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_python_functions.R
\name{grada_py_make_reverse_complement}
\alias{grada_py_make_reverse_complement}
\title{GRADA - Python}
\usage{
grada_py_make_reverse_complement(file)
}
\arguments{
\item{file}{Path to the adapter file}
}
\description{
This function will reverse complement all adapters (if they have standard bases)
}
