% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_table_functions.R
\name{grada_table}
\alias{grada_table}
\title{GRADA - table}
\usage{
grada_table()
}
\value{
a table of the found adapters (forward R1 and reverse R2) with mismatches (M)
}
\description{
This function will refer to the stdandard table function.
}
\details{
It will only accept standard settings. You cannot define anything here. Please use another _table function, if you want to change the in-/output.

Standard table function: grada_table_DT()
}
