% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_table_functions.R
\name{grada_table_DT}
\alias{grada_table_DT}
\title{GRADA - Datatable (DT)}
\usage{
grada_table_DT(input = "temp/")
}
\arguments{
\item{input}{the folder where "grada_table.txt" is saved. (std. "temp/")}
}
\value{
a table of the found adapters (forward R1 and reverse R2) with mismatches (M)
}
\description{
requires: library("DT")!
}
\details{
This function will render a table of the results from grada_analyze()
}
