% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_table_functions.R
\name{grada_table_md}
\alias{grada_table_md}
\title{GRADA - markdown table}
\usage{
grada_table_md(input = "temp/")
}
\arguments{
\item{input}{the folder where "grada_table.txt" is saved. (std. "temp/")}
}
\value{
a table of the found adapters (forward R1 and reverse R2) with mismatches (M)
}
\description{
This function will render a table of the results from grada_analyze() using markdown.
}
\details{
requires library(rmarkdown)
}
