% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_functions.R
\name{grada_analyze_positions}
\alias{grada_analyze_positions}
\title{GRADA - Analyze Positions}
\usage{
grada_analyze_positions(
  PE = TRUE,
  readlength = 150,
  input = "temp/",
  M_min = 0,
  M_max = 0,
  keepfiles = FALSE,
  numCores = detectCores()\%/\%2
)
}
\arguments{
\item{PE}{paired data? TRUE / FALSE (std. TRUE)}

\item{readlength}{longest read! for x-axis. (std. 150)}

\item{input}{input folder where the "grada_table.txt" is (output from grada_table())}

\item{M_min}{minimal mismatches (std. 0)}

\item{M_max}{maximal mismatches (std. 0)}

\item{keepfiles}{if FALSE: delete the temp files from grada_analyze (std. FALSE)}

\item{numCores}{Number of cores to use. If numCores=1 then the normal lapply function is used and the parallel package is not necessary! (std. detectCores()\%/\%2)}
}
\value{
Rdata matrix and can be used for your own plots as well.
}
\description{
This function will count the 1. positions of the sequences found in grada_table. Therefore grada_analyze needs to be run prior, with appropriate mismatches (-). Although mismatches can be analyzed, it is not recommended to use more than 3 mismatches. GRADA will give you a hint if the mismatch analysis is not possible. (A search without indels can be helpful)
}
