% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_functions.R
\name{grada_analyze_positions}
\alias{grada_analyze_positions}
\title{GRADA - Analyze Positions}
\usage{
grada_analyze_positions(
  PE = TRUE,
  readlength = 150,
  input = "temp/",
  numCores = detectCores()\%/\%2
)
}
\arguments{
\item{PE}{paired data? TRUE / FALSE (std. TRUE)}

\item{readlength}{longest read! for X-axis. (std. 150)}

\item{input}{input folder where the "grada_table.txt" is (output from grada_table())}

\item{numCores}{Number of cores to use. If numCores=1 then the normal lapply function is used and the parallel package is not necessary! (std. detectCores()\%/\%2)}
}
\value{
Rdata matrix and can be used for your own plots as well.
}
\description{
This function will count the 1. positions of the sequences found in grada_table. Therefore grada_table needs to be run prior. It will only search for mismatches = 0.
}
