% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grada_functions.R
\name{grada_plot}
\alias{grada_plot}
\title{GRADA - plot}
\usage{
grada_plot(
  PE = TRUE,
  readlength = 150,
  input = "temp/",
  numCores = detectCores()/2,
  skip = TRUE,
  plot_row = 2,
  plot_col = 2
)
}
\arguments{
\item{PE}{paired data? TRUE / FALSE (std. TRUE)}

\item{readlength}{longest read! for X-axis. (std. 150)}

\item{input}{input folder where the "grada_table.txt" is (output from grada_table())}

\item{numCores}{Number of cores to use. !!! script will use probably all cores if maxCores/2 is selected. (std. detectCores()/2)}

\item{skip}{if TRUE, it will skip plots for empty Data (if adapter is not found) so you will not get empty plots (sometimes it happens that the first plost is present anyway. this is due to a matrix R-command). (std. TRUE)}

\item{plot_row}{is the par(mfrow=c(plot_row,plot_col)) for arrenging the plots (std. 2)}

\item{plot_col}{is the par(mfrow=c(plot_row,plot_col)) for arrenging the plots (std. 2)}
}
\description{
This function will count the positions of the sequences found in grada_table. Therefore grada_table needs to be run prior. It will only search for mismatches = 0.
}
\details{
plots will be skipped if no adapters are found. If there are any difficulties with this function, may set skip to FALSE.
}
